Remove useless adapters for flink 1.13
