--- hudi-flink-datasource/hudi-flink1.17.x/src/main/java/org/apache/hudi/adapter/RateLimiterAdapter.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.hudi.adapter;
-
-import org.apache.flink.shaded.guava30.com.google.common.util.concurrent.RateLimiter;
-
-/**
- * Bridge class for shaded guava clazz {@code RateLimiter}.
- */
-public class RateLimiterAdapter {
-  private final RateLimiter rateLimiter;
-
-  private RateLimiterAdapter(double permitsPerSecond) {
-    this.rateLimiter = RateLimiter.create(permitsPerSecond);
-  }
-
-  public static RateLimiterAdapter create(double permitsPerSecond) {
-    return new RateLimiterAdapter(permitsPerSecond);
-  }
-
-  public void acquire() {
-    this.rateLimiter.acquire();
-  }
-}
